requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
cryptography==44.0.2
deprecation==2.1.0
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flask-supabase==0.2.1
frozenlist==1.6.0
gotrue==2.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
propelauth-flask==4.2.5
propelauth-py==4.2.5
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
realtime==2.4.3
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
storage3==0.11.3
StrEnum==0.4.15
stripe==12.0.1
supabase==2.15.1
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
